+++ crates/hyperswitch_domain_models/src/router_flow_types/access_token_auth.rs
@@ -1,0 +1,2 @@
+#[derive(Clone, Debug)]
+pub struct AccessTokenAuth;
