+++ crates/hyperswitch_domain_models/src/router_flow_types/dispute.rs
@@ -1,0 +1,7 @@
+#[derive(Debug, Clone)]
+pub struct Accept;
+#[derive(Debug, Clone)]
+pub struct Evidence;
+
+#[derive(Debug, Clone)]
+pub struct Defend;
