+++ crates/hyperswitch_domain_models/src/router_flow_types/files.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone)]
+pub struct Retrieve;
+
+#[derive(Debug, Clone)]
+pub struct Upload;
