+++ crates/hyperswitch_domain_models/src/router_flow_types/fraud_check.rs
@@ -1,0 +1,14 @@
+#[derive(Debug, Clone)]
+pub struct Sale;
+
+#[derive(Debug, Clone)]
+pub struct Checkout;
+
+#[derive(Debug, Clone)]
+pub struct Transaction;
+
+#[derive(Debug, Clone)]
+pub struct Fulfillment;
+
+#[derive(Debug, Clone)]
+pub struct RecordReturn;
