+++ crates/hyperswitch_domain_models/src/router_flow_types/payouts.rs
@@ -1,0 +1,20 @@
+#[derive(Debug, Clone)]
+pub struct PoCancel;
+
+#[derive(Debug, Clone)]
+pub struct PoCreate;
+
+#[derive(Debug, Clone)]
+pub struct PoEligibility;
+
+#[derive(Debug, Clone)]
+pub struct PoFulfill;
+
+#[derive(Debug, Clone)]
+pub struct PoQuote;
+
+#[derive(Debug, Clone)]
+pub struct PoRecipient;
+
+#[derive(Debug, Clone)]
+pub struct PoRecipientAccount;
