+++ crates/hyperswitch_domain_models/src/router_flow_types/refunds.rs
@@ -1,0 +1,4 @@
+#[derive(Debug, Clone)]
+pub struct Execute;
+#[derive(Debug, Clone)]
+pub struct RSync;
