+++ crates/hyperswitch_domain_models/src/router_flow_types/webhooks.rs
@@ -1,0 +1,2 @@
+#[derive(Clone, Debug)]
+pub struct VerifyWebhookSource;
