+++ crates/router/src/types/api/disputes.rs
@@ -8,6 +8,8 @@ pub struct DisputeId {
     pub dispute_id: String,
 }
 
+pub use hyperswitch_domain_models::router_flow_types::dispute::{Accept, Defend, Evidence};
+
 #[derive(Default, Debug)]
 pub struct DisputePayload {
     pub amount: String,
@@ -58,9 +60,6 @@ pub enum EvidenceType {
     UncategorizedFile,
 }
 
-#[derive(Debug, Clone)]
-pub struct Accept;
-
 pub trait AcceptDispute:
     services::ConnectorIntegration<
     Accept,
@@ -70,9 +69,6 @@ pub trait AcceptDispute:
 {
 }
 
-#[derive(Debug, Clone)]
-pub struct Evidence;
-
 pub trait SubmitEvidence:
     services::ConnectorIntegration<
     Evidence,
@@ -81,9 +77,6 @@ pub trait SubmitEvidence:
 >
 {
 }
-
-#[derive(Debug, Clone)]
-pub struct Defend;
 
 pub trait DefendDispute:
     services::ConnectorIntegration<
