+++ crates/router/src/types/api/payouts.rs
@@ -4,29 +4,11 @@ pub use api_models::payouts::{
     PayoutListFilters, PayoutListResponse, PayoutMethodData, PayoutRequest, PayoutRetrieveBody,
     PayoutRetrieveRequest, PixBankTransfer, SepaBankTransfer, Wallet as WalletPayout,
 };
+pub use hyperswitch_domain_models::router_flow_types::payouts::{
+    PoCancel, PoCreate, PoEligibility, PoFulfill, PoQuote, PoRecipient, PoRecipientAccount,
+};
 
 use crate::{services::api, types};
-
-#[derive(Debug, Clone)]
-pub struct PoCancel;
-
-#[derive(Debug, Clone)]
-pub struct PoCreate;
-
-#[derive(Debug, Clone)]
-pub struct PoEligibility;
-
-#[derive(Debug, Clone)]
-pub struct PoFulfill;
-
-#[derive(Debug, Clone)]
-pub struct PoQuote;
-
-#[derive(Debug, Clone)]
-pub struct PoRecipient;
-
-#[derive(Debug, Clone)]
-pub struct PoRecipientAccount;
 
 pub trait PayoutCancel:
     api::ConnectorIntegration<PoCancel, types::PayoutsData, types::PayoutsResponseData>
